timer_test completed and commented
timer_test will now run several times in an independent thread, each time through a Timer object will be created with an interval value of 3. If the timer reaches 3, it will break out and call the 'alarm' method, which will change a variable used in timer_test, causing it to cease and print a message.