Testing Arduino transmission, failure, and reset.
Arduino sends data over serial line 1100 times, then stops. When Pi detects transmission has ceased, it sends a reset signal to the Arduino. In this code, the loop() function waits for a 'R' transmission and activates a pin connected to the Reset port. In production, this needs to be a hard wire directly from the Pi.