SD card output works, need to flag the serial output.
Micro doesn't want to function when not connected to the computer through USB. Suspecting the Serial statements not allowing the rest of the program to execute when not connected to a serial port. Will set a flag when Serial is connected and check for the flag before all Serial statements.